fix seed script, add function to reformat data, cleanup files
