+++ client/utils/useFetchPhotos.js
@@ -11,7 +11,8 @@ const useFetchPhotos = (id) => {
 
       try {
         const response = await axios.get(
-          `http://18.144.125.169/listings/${id}/photos`
+          `http://localhost:5000/listings/${id}`
+          // `http://18.144.125.169/listings/${id}/photos`
         );
         if (response.status === 200) {
           setPhotos(response.data);
