--- config/config.js
@@ -1,2 +1,0 @@
-db: 'mongodb://exampleUser:changeThisInfo@localhost:27017/photo-service',
-db_dev: 'mongodb://exampleUser:changeThisInfo@localhost:27017/photo-service'
